\name{job.kda}
\alias{job.kda}
\docType{data}
\title{
Key Driver Analyzing results
}
\description{
Key Driver Analysis (KDA) applied data object. 
"modules.mousecoexpr.liver.human.txt" including the coexpression modules 
and "network.mouseliver.mouse.txt" including the network (graph)
information files (under the extdata folder) were used for the KDA.
}
\examples{
data(job.kda)
}
\format{
The format is: list
}
\keyword{datasets}
